Exception v2

+++ main.py
@@ -1,5 +1,7 @@
 """This is the code of my bot"""
 from datetime import timedelta, date
+from threading import Thread
+from time import sleep
 import os
 import sqlite3
 import telebot
@@ -11,7 +13,7 @@ curs1.execute('''create table if not exists birthdays(id_tel varchar2(200),
 note varchar2(1000), d_birthday date)''')
 curs1.close()
 
-#bot = telebot.TeleBot(os.environ['my_token'])
+# bot = telebot.TeleBot(os.environ['my_token'])
 bot = telebot.TeleBot('5019599335:AAFvC46wOT3vX2GK-53gqLyJwBm8yQowWZM')
 
 mp = dict()
@@ -51,16 +53,15 @@ def answer(label, message, human_id):
     elif label == 2:
         text = message.text
         curs3.execute(
-            f'delete from birthdays where id_tel = {human_id} and note = \'{text}\'')
-        curs3.execute(
             f'select * from birthdays where id_tel = {human_id} and note = \'{text}\'')
-        test = curs3.fetchall()
-        print(test)
-        if not test:
+        if not curs3.fetchall():
+            bot.send_message(human_id, 'Ну я же попросил...Ошибка: такого пользователя нет')
+        else:
+            curs3.execute(
+                f'delete from birthdays where id_tel = {human_id} and note = \'{text}\'')
             conn3.commit()
             bot.send_message(human_id, 'Я удалиль, теперь этого клоуна нет в твоих данных')
-        else:
-            bot.send_message(human_id, 'Ну я же попросил...Ошибка: такого пользователя нет')
+
     elif label == 3:
         tomorrow = (date.today() + timedelta(days=1)).strftime("%d.%m")
         if curs3.execute(
@@ -94,29 +95,42 @@ def message_reply(message):
     elif message.text == "Проверить Дни Рождения":
         mp[human_id] = 3
     elif message.text == "Вывести созданные данные":
-        if curs.execute(
-                'select note, d_birthday from birthdays where id_telegram = {hunam_id}'):
-            for lines in curs.fetchall():
+        curs.execute(
+            f'select note, d_birthday from birthdays where id_tel = {human_id}')
+        birthdays = curs.fetchall()
+        if birthdays:
+            for lines in birthdays:
                 bot.send_message(human_id, ''.join(FORM.format(line) for line in lines))
+        else:
+            bot.send_message(human_id, 'Никого неть :c')
     curs.close()
 
 
-def start():
+def start1():
     """Start out bot"""
     bot.polling(none_stop=True, interval=1)
 
 
-@bot.message_handler(content_types='text')
+def birthday_check():
+    while True:
+        check()
+        sleep(5)
+
+
 def check():
     """Check birthday"""
-    conn2 = sqlite3.connect('telegram_bot.db')
-    curs2 = conn2.cursor()
-    tomorrow = (date.today() + timedelta(days=1)).strftime("%d.%m.%Y")
-    curs2.execute(
-        f'select id_tel, note from birthdays where v_birth_day = {tomorrow}')
-    for line in curs2.fetchall():
-        bot.send_message(line[0], 'Не забудь поздравить ' + line[1] + ' с Днём Рождеия')
-    bot.polling(none_stop=True, interval=8)
+    while True:
+        conn2 = sqlite3.connect('telegram_bot.db')
+        curs2 = conn2.cursor()
+        tomorrow = (date.today() + timedelta(days=1)).strftime("%d.%m")
+        curs2.execute(
+            f'select id_tel, note from birthdays where d_birthday = {tomorrow}')
+        for line in curs2.fetchall():
+            bot.send_message(line[0], f'Не забудь поздравить {line[1]} с Днём Рождеия')
+        bot.polling(none_stop=True, interval=2)
 
 
-start()
+th = Thread(target=birthday_check())
+th.start()
+
+start1()
